frontend: bugfix
* fix a bug with the category selector:
since we use React.useState for categories in the BuildRow Submodule we never trigger the update on the main page. This means that changes to the category were never pushed to the Main Page for usage in e.g. the search function
This change updates this by tracking the state in the parent, and adding a updateXXX function
* implement the added / removed fields for the onChange function of the CategoryPicker
* fix search for categories not including subcategories